23:24

+++ Models/User.php
@@ -67,7 +67,7 @@ class User
 	//Este método selecciona todas las tuplas de la tabla
 	//usuario en caso de error se muestra por pantalla.
 
-	public function get()
+	public function get($option)
 	{
 		try
 		{
@@ -75,27 +75,36 @@ class User
 			FROM v_mostrarUsuarios";
 			//Sentencia SQL para selección de datos.
 			$stm = sqlsrv_query($this->con, $sql);
-
-			while($r = sqlsrv_fetch_array($stm)){
-				$id_usuario = $r['id_usuario'];
-				$nombre_completo = $r['nombre_completo'];
-				$correo = $r['correo'];
-				$telefono = $r['telefono'];
-				?>
-					<tr class="bg-light">
-						<td><?php echo $id_usuario; ?></td>
-						<td><?php echo $nombre_completo; ?></td>
-						<td><?php echo $correo; ?></td>
-						<td><?php echo $telefono; ?></td>
-						<td>
-							<button type="button" class="btn btn-success editbtn" data-toggle="modal" data-target="#editar">Editar</button>
-							<button type="button" class="btn btn-danger mt-0">
-								<a class="link" onclick="javascript:return confirm('¿Seguro de eliminar este registro?');" href="?c=UserController&a=delete&id=<?php echo $r['id_usuario']; ?>">Eliminar</a>
-							</button>
-						</td>
-					</tr>
-				<?php 
-		    }  
+			if($option==1){
+				while($r = sqlsrv_fetch_array($stm)){
+					$id_usuario = $r['id_usuario'];
+					$nombre_completo = $r['nombre_completo'];
+					$correo = $r['correo'];
+					$telefono = $r['telefono'];
+					?>
+						<tr class="bg-light">
+							<td><?php echo $id_usuario; ?></td>
+							<td><?php echo $nombre_completo; ?></td>
+							<td><?php echo $correo; ?></td>
+							<td><?php echo $telefono; ?></td>
+							<td>
+								<button type="button" class="btn btn-success editbtn" data-toggle="modal" data-target="#editar">Editar</button>
+								<button type="button" class="btn btn-danger mt-0">
+									<a class="link" onclick="javascript:return confirm('¿Seguro de eliminar este registro?');" href="?c=UserController&a=delete&id=<?php echo $r['id_usuario']; ?>">Eliminar</a>
+								</button>
+							</td>
+						</tr>
+					<?php 
+				}  
+			}else if($option==2){
+				while($r = sqlsrv_fetch_array($stm)){
+					$id_usuario = $r['id_usuario'];
+					$nombre_completo = $r['nombre_completo'];
+					?>
+						<option value="<?php echo $id_usuario; ?>"><?php echo $nombre_completo; ?></option>
+					<?php 
+				}
+			}
 	    }
 		catch(Exception $e)
 		{
